refactor: extract useAddUploadFilesForm from AddUploadFilesForm
423 lines and 16 functions held the whole per-file pipeline (alt-text
review, then multi-variant cropping), two mutations, drag-and-drop, and
the form JSX.

Move state and handlers into useAddUploadFilesForm, leaving the component
as JSX (423 -> 181 lines).

Pure extraction. Two details preserved deliberately:
- handleReset is declared with `function` because the useAddUploadImageSet
  onSuccess callback above it calls it, relying on hoisting. Converting to
  const arrow functions would make that a TDZ error at runtime.
- hook call order is unchanged.

Verified: form JSX byte-identical to main (128 lines, zero diff);
tsc -p tsconfig.app.json stays at 0 errors.
